moved movie class to movie-search module
